lcw
sync

+++ YiYiProject/YiYiProject/Controllers/T台/ViewControllers/LPhotoBrowser.m
@@ -342,12 +342,12 @@
 //到商场的
 -(void)turntoshangchang:(UITapGestureRecognizer *)sender{
     
-    NSLog(@"xxxshanchang==%ld",sender.view.tag);
+    NSLog(@"xxxshanchang==%ld",(long)sender.view.tag);
     
     UILabel *testlabel=(UILabel *)sender.view;
     
     GStorePinpaiViewController *detail = [[GStorePinpaiViewController alloc]init];
-    detail.storeIdStr =[NSString stringWithFormat:@"%ld",sender.view.tag] ;
+    detail.storeIdStr =[NSString stringWithFormat:@"%ld",(long)sender.view.tag] ;
     detail.storeNameStr=testlabel.text;
     [self.navigationController pushViewController:detail animated:YES];
